Add some testing to main
